added functionality to is design ready to start

+++ src/main/java/yosadchuk/needle/flow/model/dto/AddInventoryRequestDto.java
@@ -1,4 +1,6 @@
 package yosadchuk.needle.flow.model.dto;
 
-public record AddInventoryRequestDto(Integer threadId, Integer addSkeins, Double addBobbinMeters) {
+import java.math.BigDecimal;
+
+public record AddInventoryRequestDto(Integer threadId, Integer addSkeins, BigDecimal addBobbinMeters) {
 }
