+++ src/main/java/yosadchuk/needle/flow/model/dto/InventoryResponseDto.java
@@ -2,6 +2,8 @@ package yosadchuk.needle.flow.model.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
-public record InventoryResponseDto(Integer id, Integer skeinsQuantity, Double bobbinMeters, Double totalMeters) {
+public record InventoryResponseDto(Integer id, Integer skeinsQuantity, BigDecimal bobbinMeters, BigDecimal totalMeters) {
 }
